Remove superfluous response
In this case, if we can't find or access a file, the distinction is
irrelevant.  The generic error message would hint that the file actually
does exist, but is inaccessible for some other reason (insufficient
permissions, etc.). This behavior could have been leveraged by a
malicious user and should be removed.