Remove "self" variable
Previously, "self" was being used to maintain a pointer to the "this"
value outside of the anonymous function. Instead, we use the native
Function.bind(...) method to bind the method to the caller's namespace.

Beyond removing a superfluous variable, this also prevents a user from
mistakenly referencing "this" inside the anonymous function instead of
"self".